refactor(composition): fold composition into runtime
