--- CruiserJumpPractice/Composition/PluginComposition.cs
@@ -1,65 +1,0 @@
-#nullable enable
-
-using BepInEx.Logging;
-using CruiserJumpPractice.Core.Ports;
-using CruiserJumpPractice.Core.Runtime;
-using CruiserJumpPractice.Core.State;
-using CruiserJumpPractice.Core.UseCases.Client;
-using CruiserJumpPractice.Core.UseCases.Server;
-using CruiserJumpPractice.Interop;
-using CruiserJumpPractice.Interop.Game;
-using CruiserJumpPractice.Interop.InputUtils;
-
-namespace CruiserJumpPractice.Composition;
-
-internal static class PluginComposition
-{
-    public static PluginRuntime Create(ManualLogSource logger)
-    {
-        var coreLogger = new BepInExCoreLogger(logger);
-        var inputActions = new InputUtilsActions();
-        var practiceInput = new InputUtilsPracticeInput(inputActions);
-        IGameInterop gameInterop = new GameInterop(logger);
-
-        var cruiserStateStore = new CruiserStateStore();
-        var saveCruiserStateUseCase = new SaveCruiserStateUseCase(
-            gameInterop,
-            cruiserStateStore,
-            coreLogger
-        );
-        var loadCruiserStateUseCase = new LoadCruiserStateUseCase(
-            gameInterop,
-            cruiserStateStore,
-            coreLogger
-        );
-
-        var requestSaveCruiserStateUseCase = new RequestSaveCruiserStateUseCase(gameInterop);
-        var requestLoadCruiserStateUseCase = new RequestLoadCruiserStateUseCase(gameInterop);
-        var toggleMagnetUseCase = new ToggleMagnetUseCase(gameInterop);
-        var presentSaveCruiserStateResultUseCase = new PresentSaveCruiserStateResultUseCase(
-            gameInterop,
-            coreLogger
-        );
-        var presentLoadCruiserStateResultUseCase = new PresentLoadCruiserStateResultUseCase(
-            gameInterop,
-            coreLogger
-        );
-
-        var frameHandler = new FrameHandler(
-            gameInterop,
-            practiceInput,
-            requestSaveCruiserStateUseCase,
-            requestLoadCruiserStateUseCase,
-            toggleMagnetUseCase
-        );
-
-        return new PluginRuntime(
-            frameHandler: frameHandler,
-            startupHandler: new StartupHandler(gameInterop),
-            saveCruiserStateUseCase: saveCruiserStateUseCase,
-            loadCruiserStateUseCase: loadCruiserStateUseCase,
-            presentSaveCruiserStateResultUseCase: presentSaveCruiserStateResultUseCase,
-            presentLoadCruiserStateResultUseCase: presentLoadCruiserStateResultUseCase
-        );
-    }
-}
