+++ CruiserJumpPractice/Composition/PluginRuntime.cs
@@ -1,9 +1,15 @@
 #nullable enable
 
+using BepInEx.Logging;
+using CruiserJumpPractice.Core.Ports;
 using CruiserJumpPractice.Core.Runtime;
+using CruiserJumpPractice.Core.State;
 using CruiserJumpPractice.Core.UseCases;
 using CruiserJumpPractice.Core.UseCases.Client;
 using CruiserJumpPractice.Core.UseCases.Server;
+using CruiserJumpPractice.Interop;
+using CruiserJumpPractice.Interop.Game;
+using CruiserJumpPractice.Interop.InputUtils;
 
 namespace CruiserJumpPractice.Composition;
 
@@ -16,7 +22,7 @@ internal sealed class PluginRuntime
     private readonly PresentSaveCruiserStateResultUseCase presentSaveCruiserStateResultUseCase;
     private readonly PresentLoadCruiserStateResultUseCase presentLoadCruiserStateResultUseCase;
 
-    public PluginRuntime(
+    private PluginRuntime(
         FrameHandler frameHandler,
         StartupHandler startupHandler,
         SaveCruiserStateUseCase saveCruiserStateUseCase,
@@ -31,6 +37,55 @@ internal sealed class PluginRuntime
         this.loadCruiserStateUseCase = loadCruiserStateUseCase;
         this.presentSaveCruiserStateResultUseCase = presentSaveCruiserStateResultUseCase;
         this.presentLoadCruiserStateResultUseCase = presentLoadCruiserStateResultUseCase;
+    }
+
+    public static PluginRuntime Create(ManualLogSource logger)
+    {
+        var coreLogger = new BepInExCoreLogger(logger);
+        var inputActions = new InputUtilsActions();
+        var practiceInput = new InputUtilsPracticeInput(inputActions);
+        IGameInterop gameInterop = new GameInterop(logger);
+
+        var cruiserStateStore = new CruiserStateStore();
+        var saveCruiserStateUseCase = new SaveCruiserStateUseCase(
+            gameInterop,
+            cruiserStateStore,
+            coreLogger
+        );
+        var loadCruiserStateUseCase = new LoadCruiserStateUseCase(
+            gameInterop,
+            cruiserStateStore,
+            coreLogger
+        );
+
+        var requestSaveCruiserStateUseCase = new RequestSaveCruiserStateUseCase(gameInterop);
+        var requestLoadCruiserStateUseCase = new RequestLoadCruiserStateUseCase(gameInterop);
+        var toggleMagnetUseCase = new ToggleMagnetUseCase(gameInterop);
+        var presentSaveCruiserStateResultUseCase = new PresentSaveCruiserStateResultUseCase(
+            gameInterop,
+            coreLogger
+        );
+        var presentLoadCruiserStateResultUseCase = new PresentLoadCruiserStateResultUseCase(
+            gameInterop,
+            coreLogger
+        );
+
+        var frameHandler = new FrameHandler(
+            gameInterop,
+            practiceInput,
+            requestSaveCruiserStateUseCase,
+            requestLoadCruiserStateUseCase,
+            toggleMagnetUseCase
+        );
+
+        return new PluginRuntime(
+            frameHandler: frameHandler,
+            startupHandler: new StartupHandler(gameInterop),
+            saveCruiserStateUseCase: saveCruiserStateUseCase,
+            loadCruiserStateUseCase: loadCruiserStateUseCase,
+            presentSaveCruiserStateResultUseCase: presentSaveCruiserStateResultUseCase,
+            presentLoadCruiserStateResultUseCase: presentLoadCruiserStateResultUseCase
+        );
     }
 
     public void HandleStartup()
